+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 
 - **Breaking change**: Refactored the [HexaPDF::Document] interface for working
   with objects and move parts into [HexaPDF::Revisions]
+- **Breaking change**: [HexaPDF::Layout::TextBox] to use whole available width
+  when aligning to the center or right
 - CLI command `hexapdf inspect` to show the type of revisions, as well as the
   number of objects per revision